refactor: HTTPレスポンスのlinkヘッダーから次のページのURLを取得する処理をClient内から切り出す
