+++ GitHubClientApp/UserRepository/UserRepositoryClient.swift
@@ -30,25 +30,6 @@ struct UserRepositoryClient {
         let repositories: [Repository] = try ResponseTranslator.from(response: response)
         
         
-        return (repositories, extractNextPageLink(response.headers))
-    }
-    
-    private func extractNextPageLink(_ headers: [String: String]) -> String? {
-        guard let headerValue = headers["link"], headerValue.contains("rel=\"next\"") else {
-            return nil
-        }
-        
-        let links = headerValue
-            .split(separator: ",")
-            .filter { $0.contains("rel=\"next\"")}
-        
-        let pattern = /<([^>]*)>/
-        
-        guard let link = links.first, let match = link.firstMatch(of: pattern) else {
-            return nil
-        }
-        
-        return String(match.1)
-        
+        return (repositories, NextPageLinkExtractor.getNextPageLink(from: response.headers))
     }
 }
